backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
django-auditlog==3.0.0
django-storages==1.14.4
django-redis==5.4.0
django-celery-beat==2.6.0
celery==5.4.0
redis==5.0.7
Pillow==10.4.0
WeasyPrint==62.3
openpyxl==3.1.5
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.2.0
drf-spectacular==0.27.2
django-guardian==2.4.0
boto3==1.35.0
requests==2.32.3
phonenumbers==8.13.45
anthropic==0.34.2
openai==1.51.0
whitenoise==6.7.0
gunicorn==23.0.0
